ingest-car-data/requirements.txt
google-cloud-bigquery
google-cloud-secret-manager
git+https://github.com/tillsteinbach/CarConnectivity.git
git+https://github.com/tillsteinbach/CarConnectivity-connector-volkswagen.git
